# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients/{id}": {
      "get": {
        "operationId": "PatientsController_findOne",
        "summary": "Buscar paciente por ID",
        "description": "Retorna os detalhes de um paciente específico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "summary": "Atualizar paciente",
        "description": "Atualiza as informações de um paciente existente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paciente atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos fornecidos"
          },
          "404": {
            "description": "Paciente não encontrado"
          },
          "409": {
            "description": "Conflito com outro paciente (CPF ou email já existente)"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PatientsController_remove",
        "summary": "Remover paciente",
        "description": "Remove um paciente do sistema",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Paciente removido com sucesso"
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "patients"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/patients/{patientId}/photos": {
      "post": {
        "operationId": "PatientPhotosController_create",
        "summary": "Create a new patient photo",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientPhotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The photo has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPhoto"
                }
              }
            }
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PatientPhotosController_findAll",
        "summary": "Get all photos for a patient",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patient photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPhoto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/patients/{patientId}/photos/type/{type}": {
      "get": {
        "operationId": "PatientPhotosController_findByType",
        "summary": "Get photos by type for a patient",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patient photos of a specific type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPhoto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/patients/{patientId}/photos/date-range": {
      "get": {
        "operationId": "PatientPhotosController_findByDateRange",
        "summary": "Get photos within a date range",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of patient photos within the specified date range",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPhoto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/patients/{patientId}/photos/{id}": {
      "get": {
        "operationId": "PatientPhotosController_findOne",
        "summary": "Get a specific photo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found photo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPhoto"
                }
              }
            }
          },
          "404": {
            "description": "Photo not found"
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "PatientPhotosController_update",
        "summary": "Update a photo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientPhotoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The photo has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPhoto"
                }
              }
            }
          },
          "404": {
            "description": "Photo not found"
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PatientPhotosController_remove",
        "summary": "Delete a photo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The photo has been successfully deleted."
          },
          "404": {
            "description": "Photo not found"
          }
        },
        "tags": [
          "patient-photos"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/meal-plans/patient/{patientId}": {
      "get": {
        "operationId": "MealPlansController_findByPatient",
        "summary": "Buscar planos por paciente",
        "description": "Retorna todos os planos alimentares de um paciente específico",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "description": "ID do paciente",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de planos alimentares retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealPlan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Paciente não encontrado"
          }
        },
        "tags": [
          "meal-plans"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartNutri API",
    "description": "API para gestão completa de nutrição",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do paciente"
          },
          "name": {
            "type": "string",
            "description": "Nome do paciente"
          },
          "email": {
            "type": "string",
            "description": "Email do paciente"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do paciente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do paciente"
          },
          "address": {
            "type": "string",
            "description": "Endereço do paciente"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "cpf",
          "phone",
          "address",
          "nutritionistId"
        ]
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do paciente"
          },
          "email": {
            "type": "string",
            "description": "Email do paciente"
          },
          "cpf": {
            "type": "string",
            "description": "CPF do paciente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do paciente"
          },
          "address": {
            "type": "string",
            "description": "Endereço do paciente"
          }
        }
      },
...
      "CreatePatientPhotoDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "description": "ID do paciente",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "photoUrl": {
            "type": "string",
            "description": "URL da foto",
            "example": "https://storage.example.com/photos/patient123-front.jpg"
          },
          "photoType": {
            "type": "string",
            "description": "Tipo da foto",
            "enum": [
              "front",
              "back",
              "left_side",
              "right_side",
              "other"
            ],
            "example": "front"
          },
          "photoDate": {
            "format": "date-time",
            "type": "string",
            "description": "Data da foto",
            "example": "2024-04-01"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre a foto",
            "example": "Foto tirada após 3 meses de dieta"
          }
        },
        "required": [
          "patientId",
          "photoUrl",
          "photoType",
          "photoDate"
        ]
      },
      "PatientPhoto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID da foto"
          },
          "patientId": {
            "type": "string",
            "description": "ID do paciente"
          },
          "nutritionistId": {
            "type": "string",
            "description": "ID do nutricionista"
          },
          "photoUrl": {
            "type": "string",
            "description": "URL da foto"
          },
          "photoType": {
            "type": "string",
            "description": "Tipo da foto",
            "enum": [
              "front",
              "back",
              "left_side",
              "right_side",
              "other"
            ],
            "example": "front"
          },
          "photoDate": {
            "format": "date-time",
            "type": "string",
            "description": "Data da foto"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre a foto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação do registro"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data da última atualização do registro"
          }
        },
        "required": [
          "id",
          "patientId",
          "nutritionistId",
          "photoUrl",
          "photoType",
          "photoDate",
          "notes",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePatientPhotoDto": {
        "type": "object",
        "properties": {
          "photoUrl": {
            "type": "string",
            "description": "URL da foto",
            "example": "https://storage.example.com/photos/patient123-front-updated.jpg"
          },
          "photoType": {
            "type": "string",
            "description": "Tipo da foto",
            "enum": [
              "front",
              "back",
              "left_side",
              "right_side",
              "other"
            ],
            "example": "front"
          },
          "photoDate": {
            "format": "date-time",
            "type": "string",
            "description": "Data da foto",
            "example": "2024-04-01"
          },
          "notes": {
            "type": "string",
            "description": "Observações sobre a foto",
            "example": "Foto tirada após 3 meses de dieta"
          }
        }
      },
...
      "Nutritionist": {
        "type": "object",
        "properties": {}
      },
      "Meal": {
        "type": "object",
        "properties": {
          "totalCalories": {
            "type": "number",
            "example": 500.25
          },
          "totalProtein": {
            "type": "number",
            "example": 30.5
          },
          "totalCarbs": {
            "type": "number",
            "example": 60.75
          },
          "totalFat": {
            "type": "number",
            "example": 20.3
          }
        },
        "required": [
          "totalCalories",
          "totalProtein",
          "totalCarbs",
          "totalFat"
        ]
      },
      "MealPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "Plano Alimentar Semanal"
          },
          "notes": {
            "type": "string",
            "example": "Plano alimentar para perda de peso"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-27"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "patientId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nutritionist": {
            "$ref": "#/components/schemas/Nutritionist"
          },
          "nutritionistId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20T10:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-03-20T10:00:00Z"
          },
          "dailyCalories": {
            "type": "number",
            "example": 2000.5
          },
          "dailyProtein": {
            "type": "number",
            "example": 150.75
          },
          "dailyCarbs": {
            "type": "number",
            "example": 250.3
          },
          "dailyFat": {
            "type": "number",
            "example": 80.2
          }
        },
        "required": [
          "id",
          "name",
          "notes",
          "startDate",
          "endDate",
          "patient",
          "patientId",
          "nutritionist",
          "nutritionistId",
          "meals",
          "createdAt",
          "updatedAt",
          "dailyCalories",
          "dailyProtein",
          "dailyCarbs",
          "dailyFat"
        ]
      },
...
    }
  }
}